refactor : create UpdateNightBoothReservationBean and use UpdateNightBoothReservationBean when modifying reservation availability
